Refactored db_model_to_game_model to dependencies.py
Things still work
