+++ civ_vi_webhook/api/info_endpoints.py
@@ -3,7 +3,8 @@ from typing import Optional
 import fastapi.responses
 from fastapi import APIRouter, Query
 
-from ..models.api import information_models, games
+from ..dependencies import db_model_to_game_model
+from ..models.api import information_models
 from ..services.db import game_service, user_service
 
 router = APIRouter(tags=['Information Endpoints'])
@@ -38,26 +39,6 @@ async def return_completed_games():
     games_to_return = await db_model_to_game_model(completed_games)
     print(games_to_return)
     return {"games": games_to_return}
-
-
-async def db_model_to_game_model(these_games):
-    games_to_return = []
-    for game in these_games:
-        time_stamp = games.TimeStamp(year=game.game_info.time_stamp.year,
-                                     month=game.game_info.time_stamp.month,
-                                     day=game.game_info.time_stamp.day,
-                                     hour=game.game_info.time_stamp.hour,
-                                     minute=game.game_info.time_stamp.minute,
-                                     second=game.game_info.time_stamp.second)
-        player_name = await user_service.get_index_name_by_user_id(game.game_info.next_player_id)
-        game_info = games.GameInfo(player_name=player_name, turn_number=game.game_info.turn_number,
-                                   game_completed=game.game_info.game_completed, time_stamp=time_stamp,
-                                   turn_deltas=game.game_info.turn_deltas,
-                                   average_turn_time=game.game_info.average_turn_time,
-                                   winner=game.game_info.winner)
-        this_game = games.Game(game_name=game.game_name, game_info=game_info)
-        games_to_return.append(this_game)
-    return games_to_return
 
 
 @router.get('/all_games', response_model=information_models.CurrentGames)
